+++ NEWS.md
@@ -1,6 +1,11 @@
 NEWS
 ===========
 
+### rtorf 2.8.1 (Release date: 2025-10-15)
+
+Remove need for columns site_code or altitude_final in obs_agg
+User may need to reorder data
+
 ### rtorf 2.8.0 (Release date: 2025-10-15)
 
 add obs_select_sec to select based on closest time, rather than agg